Moved music related tests to a separate file.
