+++ src/music.ts
@@ -124,7 +124,7 @@ export abstract class AbstractNote {
  * Represents an alphabetic musical note (A-G).
  */
 export class MusicNote extends AbstractNote {
-    public static readonly PATTERN = /^([A-G])(♮|#|♯|b|♭|[ei]s|s)?/;
+    public static readonly PATTERN = /^([A-G])(♮|#|♯|b|♭|[ei]s|s)?$/;
 
     /**
      * Create a new MusicNote instance.
@@ -151,7 +151,7 @@ export class MusicNote extends AbstractNote {
         }
 
         const root = match[1].toUpperCase();
-        const postfix = match[2] ? match[2].toLowerCase() : undefined;
+        const postfix = match[2];
 
         return new MusicNote(root, postfix);
     }
@@ -161,7 +161,7 @@ export class MusicNote extends AbstractNote {
  * Represents a Nashville number notation (1-7).
  */
 export class NashvilleNumber extends AbstractNote {
-    public static readonly PATTERN = /^([1-7])(#|♯|b|♭)?/;
+    public static readonly PATTERN = /^([1-7])(#|♯|b|♭)?$/;
 
     /**
      * Create a new NashvilleNumber instance.
@@ -195,7 +195,7 @@ export class NashvilleNumber extends AbstractNote {
         }
 
         const root = match[1];
-        const postfix = match[2] ? match[2].toLowerCase() : undefined;
+        const postfix = match[2];
 
         return new NashvilleNumber(parseInt(root), postfix);
     }
