+++ test/transpose.test.ts
@@ -10,8 +10,6 @@ import {
     AbstractNote, 
     Accidental,
     MusicNote,
-    NashvilleNumber,
-    MusicTheory,
     KeyInfo,
     AbsoluteKeyInfo,
     MajorKeyInfo,
@@ -27,168 +25,7 @@ import {
 
 import { verifyChordsInBlock } from "./parser.test";
 
-describe("MusicTheory", () => {
-    describe("getNoteIndex", () => {
-        const noteIndexTestCases = [
-            { pitch: "C", accidental: undefined, expectedIndex: 0 },
-            { pitch: "C", accidental: "#", expectedIndex: 1 },
-            { pitch: "D", accidental: "b", expectedIndex: 1 },
-            { pitch: "D", accidental: undefined, expectedIndex: 2 },
-            { pitch: "F", accidental: "#", expectedIndex: 6 },
-            { pitch: "G", accidental: "b", expectedIndex: 6 },
-            { pitch: "B", accidental: undefined, expectedIndex: 11 },
-        ];
 
-        test.each(noteIndexTestCases)(
-            "should return chromatic index $expectedIndex for $pitch$accidental",
-            ({ pitch, accidental, expectedIndex }) => {
-                const note = new MusicNote(pitch, accidental);
-                const noteIndex = MusicTheory.getNoteIndex(note);
-                expect(noteIndex).toEqual(expectedIndex);
-            }
-        );
-    });
-
-    describe("getPreferredNoteName", () => {
-        const preferredNoteTestCases = [
-            // Natural notes (single options)
-            { index: 0, preference: undefined, expected: "C" },
-            { index: 2, preference: undefined, expected: "D" },
-            { index: 4, preference: undefined, expected: "E" },
-            { index: 5, preference: undefined, expected: "F" },
-            { index: 7, preference: undefined, expected: "G" },
-            { index: 9, preference: undefined, expected: "A" },
-            { index: 11, preference: undefined, expected: "B" },
-
-            // Sharp preference
-            { index: 1, preference: Accidental.SHARP, expected: "C#" },
-            { index: 3, preference: Accidental.SHARP, expected: "D#" },
-            { index: 6, preference: Accidental.SHARP, expected: "F#" },
-            { index: 8, preference: Accidental.SHARP, expected: "G#" },
-            { index: 10, preference: Accidental.SHARP, expected: "A#" },
-
-            // Flat preference
-            { index: 1, preference: Accidental.FLAT, expected: "Db" },
-            { index: 3, preference: Accidental.FLAT, expected: "Eb" },
-            { index: 6, preference: Accidental.FLAT, expected: "Gb" },
-            { index: 8, preference: Accidental.FLAT, expected: "Ab" },
-            { index: 10, preference: Accidental.FLAT, expected: "Bb" },
-        ];
-
-        test.each(preferredNoteTestCases)(
-            "should return $expected for index $index with preference $preference",
-            ({ index, preference, expected }) => {
-                expect(
-                    MusicTheory.getPreferredNoteName(index, preference)
-                ).toEqual(expected);
-            }
-        );
-    });
-});
-
-describe('KeyInfo', () => {
-    describe('parse major keys', () => {
-        const majorKeyTestCases = [
-            { input: 'C', root: 'C', accidental: Accidental.NATURAL, minor: 'Am' },
-            { input: 'G', root: 'G', accidental: Accidental.NATURAL, minor: 'Em' },
-            { input: 'F#', root: 'F#', accidental: Accidental.SHARP, minor: 'D#m' },
-            { input: 'Bb', root: 'Bb', accidental: Accidental.FLAT, minor: 'Gm' },
-            { input: 'Db', root: 'Db', accidental: Accidental.FLAT, minor: 'Bbm' },
-            { input: 'C#', root: 'C#', accidental: Accidental.SHARP, minor: 'A#m' },
-        ];
-
-        test.each(majorKeyTestCases)(
-            'should parse $input correctly',
-            ({ input, root, accidental, minor }) => {
-                const key = MajorKeyInfo.parse(input);
-
-                expect(key).toBeInstanceOf(MajorKeyInfo);
-                expect(key.toString()).toEqual(input);
-
-                expect(key.root.toString()).toEqual(root);
-                expect(key.accidental).toEqual(accidental);
-
-                // verify major key info properties
-                expect(key.getScaleDegrees()).toEqual([0, 2, 4, 5, 7, 9, 11]);
-
-                // verify relative minor key
-                const expectedMinor = KeyInfo.parse(minor) as MinorKeyInfo;
-                const relativeMinor = key.getRelativeMinor();
-                expect(relativeMinor).toEqual(expectedMinor);
-            }
-        );
-
-    });
-
-    describe('parse minor keys', () => {
-        const minorKeyTestCases = [
-            { input: 'Am', root: 'A', accidental: Accidental.NATURAL, major: 'C' },
-            { input: 'Em', root: 'E', accidental: Accidental.NATURAL, major: 'G' },
-            { input: 'F#m', root: 'F#', accidental: Accidental.SHARP, major: 'A' },
-            { input: 'Bbm', root: 'Bb', accidental: Accidental.FLAT, major: 'Db' },
-            { input: 'C#m', root: 'C#', accidental: Accidental.SHARP, major: 'E' },
-            { input: 'Ebm', root: 'Eb', accidental: Accidental.FLAT, major: 'Gb' },
-        ];
-
-        test.each(minorKeyTestCases)(
-            'should parse $input correctly',
-            ({ input, root, accidental, major }) => {
-                const key = MinorKeyInfo.parse(input);
-
-                expect(key).toBeInstanceOf(MinorKeyInfo);
-                expect(key.toString()).toEqual(input);
-
-                expect(key.root.toString()).toEqual(root);
-                expect(key.accidental).toEqual(accidental);
-
-                // verify minor key info properties
-                expect(key.getScaleDegrees()).toEqual([0, 2, 3, 5, 7, 8, 10]);
-
-                // verify relative major key
-                const expectedMajor = KeyInfo.parse(major) as MajorKeyInfo;
-                const relativeMajor = key.getRelativeMajor();
-                expect(relativeMajor).toEqual(expectedMajor);
-            }
-        );
-
-        const invalidKeyTestCases = [
-            'H',
-            'C#b',
-            '',
-            'CM',
-            'Am#',
-            'X',
-            '1',
-        ];
-
-        test.each(invalidKeyTestCases)(
-            'should throw on invalid key %s',
-            (invalidKey) => {
-                expect(() => KeyInfo.parse(invalidKey)).toThrow();
-            }
-        );
-    });
-
-    describe('getInterval', () => {
-        const intervalTestCases = [
-            { from: new MusicNote('C'), to: new MusicNote('C'), expected: 0 },
-            { from: new MusicNote('C'), to: new MusicNote('D'), expected: 2 },
-            { from: new MusicNote('C'), to: new MusicNote('G'), expected: 7 },
-            { from: new MusicNote('C'), to: new MusicNote('C', '#'), expected: 1 },
-            { from: new MusicNote('D'), to: new MusicNote('C'), expected: 10 },
-            { from: new MusicNote('G'), to: new MusicNote('F'), expected: 10 },
-            { from: new MusicNote('A'), to: new MusicNote('A'), expected: 0 },
-            { from: new MusicNote('F', '#'), to: new MusicNote('G', 'b'), expected: 0 },
-        ];
-
-        test.each(intervalTestCases)(
-            'should calculate interval from $from.root$from.postfix to $to.root$to.postfix as $expected',
-            ({ from, to, expected }) => {
-                expect(MusicTheory.getInterval(from, to)).toEqual(expected);
-            }
-        );
-    });
-});
 
 describe('NoteTransposer', () => {
     describe('transposeNote', () => {
